Verify travel advice email alerts.
These work slightly differently from drug alerts, in that there is an existing
list of countries and alerts are sent out when any of them are updated. So
instead of checking for the existence of a link in the received emails, we
query on the update time in the body of the email and the country name in the
subject.

We also restrict the query to updates made within the last two days, since some
countries are updated very infrequently.