Exit with code 2 when missing emails found.
This allows us to distinguish between abnormal exit, for instance when the API
couldn't be contacted, which should not trigger an out-of-hours alert; and
when we have confirmed that emails are actually missing, which should.

See https://nagios-plugins.org/doc/guidelines.html#AEN78 for how Nagios
uses the return codes.