Allow basic acknowledging of specific missing emails.
We need a mechanism to prevent 2nd line from being
alerted about missing emails which have been
investigated and triaged.  If we don't, 2nd
line support may miss other undelivered emails
unless they manually check the job every hour to
make sure a new email is not on this list.

This is a basic manual process for adding exclusions.