Use the Content Store to fetch travel advice
This monitoring tool was still using the Content API to fetch data on
the Foreign Travel Advice page and Countries. We no longer support the
Content API, so we should be using the Content Store instead.

This commit changes the URL to be the Content Store URL and also fetches
the expected information in the correct place.