Switch email alert tests to be based around title
The subject of an email is based upon a title field which the travel
advice publisher has access to edit. For the most part these are
consistent with the names of the countries however they need not be.

Where the subject originates:
https://github.com/alphagov/travel-advice-publisher/blob/master/app/presenters/email_alert_presenter.rb#L28-L30

This has led to some confusing problems where we thought diacritics were
being stripped out of emails as the ones for "São Tomé and Principe"
were given the subject of "Sao Tome & Principe travel advice".

This commit changes what is searched for in the mailbox to check that it
is the title attribute rather than the country name - which will allow
for subtle differences. Ideally we will deprecate the use of that title
field as it seems to be widely not used and when it is used causes
problems like this.