Match drug alerts based on subject rather than URL
Since the migration from Govdelivery to the Notify based Email Alert API
system we no longer have URLs in the body content of an email.

To try make the query somewhat specific I've set it to check the subject
line of emails. This should be fine for the vast majority of cases but
I'm concerned that some unusual characters could cause problems.