Add a rake tasks to regenerate OAuth credentials
This commit adds two new rake tasks that together allow for the easy regeneration of the Google OAuth credentials used to log in to the email account that is checked for email alerts. There is also documentation for how to run this process.

As a bonus, linting is now available via a rake task and is run by default after the tests.