Replace acknowledged URLs with new format
This removes the URLs as they no-longer work for acknowledgements (and
presumably those documents were published long ago) and instead uses the
search format that goes into Gmail.

This took me a few goes to get right as this class is used by drug
alerts and travel alerts that have different formats so my attempts to
shield the formatting from the ACKNOWLEDGED_EMAIL_CONTENTS constant
failed.