Ignore alert for re-titled Imatinib drug alert
On 2018-09-21, a drug alert was published with the title "Imatinib
400mg Capsules (3 x 10) PL 36390/0180 : Company-led recall" and an
email was sent out.

On 2018-09-27, the title was changed to "Imatinib 400mg Capsules (3 x
10) PL 36390/0180 : Company-led Drug Alert", which caused the email
check to fail.

As the original email was sent out, and this was only a minor change,
the alert can be ignored.

```
irb(main):004:0> d = Document.find_by(content_id: 'd91f11cf-5e0d-4571-91d0-9552ca6cafe6').editions.pluck(:title, :state)
=> [["Imatinib 400mg Capsules (3 x 10)  PL 36390/0180 : Company-led Drug Alert", "published"], ["Imatinib 400mg Capsules (3 x 10)  PL 36390/0180 : Company-led recall", "superseded"], ["Imatinib 400mg Capsules (3 x 10)  PL 36390/0180 : Company-led recall", "superseded"]]
```