Remove South Sudan override
A major update for South Sudan travel advice was published with a
title of `South Sudan`. This was received by Email Alert Api and
subsequenty processed. A minor update was published from Travel Advice
Publisher and the title was changed to `South Sudan travel advice`.

When the monitoring job was called, it searched the courtesy copies
email group for an email with a subject that included the words
`South Sudan travel advice`. The search for this email failed because
the actual email sent had a title that only included `South Sudan`.

The exception was put in place to stop the alert and resolve the
incident. It is no longer needed because we only test for emails that
have been sent out up to 2 days ago.