Update email-alert-monitoring documentation
Updates the documentation for email-alert-monitoring and consolidates
the two pages under /docs into the README. I was originally only going
to move `Testing the alert checks` into the README to make it more
visible as it fit quite nicely but this left only one doc in the /docs
directory so I've shifted `Regenerating Google OAuth credentials` too.
The README is quite short I think this is fine.